+++ ts_control/src/tokio/mod.rs
@@ -3,11 +3,13 @@ use connect::connect;
 pub use connect::{
     CONTROL_PROTOCOL_VERSION, fetch_control_key, read_challenge_packet, upgrade_ts2021,
 };
+pub use id_token::{IdTokenError, fetch_id_token};
 pub use map_stream::{FilterUpdate, PeerUpdate, StateUpdate};
 use register::register;
 
 mod client;
 mod connect;
+mod id_token;
 mod map_stream;
 mod ping;
 mod prefixed_reader;